Remove dependency on dotenv gem
This commit removes the requirement to use `dotenv`. This gem was being
used to load `.env` and raise a load error if `MAILERSEND_API_TOKEN` was
not defined. That implementation can make it slightly more
complicated to load this gem when a project uses `dotenv-rails`, because
the correct `MAILERSEND_API_TOKEN` might be defined in one of the other
environment-specific `.env*` files.

To give projects more flexibility when and how `MAILERSEND_API_TOKEN` is
defined, `Mailersend::Client` will now raise an error at runtime instead
of at load when no API token is supplied.